DEV-83 refactor: searchService.save 메소드를 postService.save()내부 로 이동시켜 서비스간 참조를 막음.
